// lib/types/DecodeResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DecodeError } from "./DecodeError";
import type { Decoded } from "./Decoded";

export type DecodeResult = { "status": "ok", decoded: Decoded, } | { "status": "error", error: DecodeError, message: string, } | { "status": "invalid_context", message: string, };